update based on comments

+++ crates/core/src/protocols/gkr_exp/tests.rs
@@ -26,7 +26,7 @@ type F = BinaryField128bPolyval;
 type P = PackedBinaryPolyval1x128b;
 
 fn generate_claim_witness<'a>(
-	exponents_in_each_row: &[u128],
+	exponents_in_each_row: &[u64],
 	exponent_bit_width: usize,
 	base: Option<MultilinearWitness<'a, P>>,
 	eval_point: &[F],
@@ -100,10 +100,10 @@ fn generate_mul_witnesses_claims<'a, const LOG_SIZE: usize, const COLUMN_LEN: us
 	let mut rng = thread_rng();
 
 	let a: Vec<_> = (0..COLUMN_LEN)
-		.map(|_| rng.gen::<u128>() % (1 << exponent_bit_width))
+		.map(|_| rng.gen::<u64>() % (1 << exponent_bit_width))
 		.collect();
 	let b: Vec<_> = (0..COLUMN_LEN)
-		.map(|_| rng.gen::<u128>() % (1 << exponent_bit_width))
+		.map(|_| rng.gen::<u64>() % (1 << exponent_bit_width))
 		.collect();
 	let c: Vec<_> = a.iter().zip(&b).map(|(ai, bi)| ai * bi).collect();
 
@@ -163,8 +163,8 @@ fn witness_gen_happens_correctly() {
 
 	let mut rng = thread_rng();
 
-	let a: Vec<_> = (0..COLUMN_LEN).map(|_| rng.gen::<u8>() as u128).collect();
-	let b: Vec<_> = (0..COLUMN_LEN).map(|_| rng.gen::<u8>() as u128).collect();
+	let a: Vec<_> = (0..COLUMN_LEN).map(|_| rng.gen::<u8>() as u64).collect();
+	let b: Vec<_> = (0..COLUMN_LEN).map(|_| rng.gen::<u8>() as u64).collect();
 	let c: Vec<_> = a.iter().zip(&b).map(|(ai, bi)| ai * bi).collect();
 
 	let eval_point = [F::default(); LOG_SIZE].map(|_| <F as Field>::random(&mut rng));
